feat(tracer): Calculate gas used by EVM opcodes (#223)
## What ❔

ditto

## Why ❔

This field is expected in opcodes trace

## Apologies 

This is really dirty and I'm ashamed a bit. But this is only a reference
implementation (PoC) and our current design doesn't allow to handle it
in a proper way. We **really** should move charging for reading callee
account inside the EE.

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [X] Code has been formatted.