test: fixes for tester to run Blockchain tests (#268)
## What ❔

Run a new group of fixture-based tests (blockchain tests), covering some
multi-tx cases. We use a new evm tester implementation for that. The old
one is used as well just in case, since we are not 100% sure that all
legacy state tests are migrated properly to fixture tests.

## Why ❔

<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- The `Why` has to be clear to non-Matter Labs entities running their
own ZK Chain -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Is this a breaking change?
- [ ] Yes
- [ ] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted.

---------

Co-authored-by: Vladislav Volosnikov <Volosnikov.apmath@gmail.com>