feat: Publish binaries only on releases (#315)
## What ❔

Do not rebuild binaries for every commit in PRs. Binaries are built and
attached to new releases. Later on we can replace it with proper
release-please workflow.

Old binaries build workflow can be triggered manually if needed.

## Why ❔

Most of the time we update binaries together with the release. Updates
on every PR can be helpful in theory, but in practice they just create
an headache on every merge/rebase. And becuase our history in general is
not linear in branches, binaries in commits are broken most of the time.

It is still possible to update binaries manually, if needed.

Bonus: since we don't have binaries in the tree, it will be easier to
catch issues in CI (e.g. no outdated binaries etc.)

## Is this a breaking change?
- [ ] Yes
- [X] No

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted.